json-file: move to json/util
