+++ src/meson.build
@@ -693,8 +693,8 @@ common_sources = [
   'trx/gfx/screenshot.c',
   'trx/json/base.c',
   'trx/json/parse.c',
+  'trx/json/util/file.c',
   'trx/json/write.c',
-  'trx/json_file.c',
   'trx/log.c',
   'trx/memory.c',
   'trx/screenshot.c',
