+++ src/trx/game/lara/skin/storage.c
@@ -7,7 +7,7 @@
 #include <trx/game/game_string.h>
 #include <trx/game/lara.h>
 #include <trx/game/shell.h>
-#include <trx/json_file.h>
+#include <trx/json/util/file.h>
 #include <trx/memory.h>
 #include <trx/vector.h>
