+++ src/trx/game/stats/init.c
@@ -14,7 +14,7 @@
 #include <trx/game/shell.h>
 #include <trx/game/stats.h>
 #include <trx/json.h>
-#include <trx/json_file.h>
+#include <trx/json/util/file.h>
 #include <trx/log.h>
 #include <trx/memory.h>
 #include <trx/strings.h>
