+++ src/trx/json/util/file.c
@@ -1,4 +1,4 @@
-#include <trx/json_file.h>
+#include <trx/json/util/file.h>
 
 #include <trx/filesystem.h>
 #include <trx/game/shell.h>
